+++ lib/minitest-activemodel/matchers/validation_matcher.rb
@@ -40,6 +40,12 @@ module MiniTest
             @result = false
           end
         end
+
+        def validate_invalid_options! *options
+          if options.all? &:nil?
+            raise ArgumentError, 'You have to supply an option for this matcher'
+          end
+        end
       end
     end
   end