+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 + Add `validate_format_of` matcher.
 + Raise `ArgumentError` if no option is given for `validate_length_of` matcher.
 + Add `to_allow` and `to_not_allow` options to `ValidateFormatMatcher`.
++ Raise `ArgumentError` if no option is given for `validate_format_of` matcher.